requirements.txt
av
pedalboard==0.7.7
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cu118
torchcrepe==0.0.20
ffmpeg-python>=0.2.0
faiss-cpu==1.7.3
praat-parselmouth>=0.4.2
pyworld==0.3.4
fairseq==0.12.2
pydub==0.25.1
python-dotenv
wget
